pyproject.toml
[project]
name = "geo-analisis"
version = "0.1.0"
description = "Mapa de Viabilidad de Negocios con análisis geoespacial"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "fpdf2>=2.7.0",
    "python-dotenv>=1.0.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.23.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "geoalchemy2>=0.14.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4"]
ignore = []

[tool.ruff.isort]
known-first-party = ["app"]
